% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dragUI.R
\name{dragUI}
\alias{dragUI}
\title{dragUI}
\usage{
dragUI(id, ..., style = NULL, class = "dragelement")
}
\arguments{
\item{id}{The div id of the element}

\item{style}{The css styling}

\item{class}{The css class style}
}
\value{
dragable div element
export
}
\description{
dragUI
}

