% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dropUI.R
\name{dropUI}
\alias{dropUI}
\title{dropUI}
\usage{
dropUI(id, style = NULL, class = "dropelement", vertical = FALSE)
}
\arguments{
\item{id}{The div id of the element}

\item{style}{The css styling}

\item{class}{The css class style}

\item{vertical}{Make the drop element allow multiple drag elements vertically arranged}
}
\value{
div element with nested elements if vertical is true
export
}
\description{
dropUI
}

