% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/registerInputHandler.R, R/shinyDND.R
\docType{package}
\name{shinyDND}
\alias{shinyDND}
\alias{shinyDND-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.

The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\section{Foo functions}{

The foo functions ...
}

