% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dragSetUI.R
\name{dragSetUI}
\alias{dragSetUI}
\title{dragSetUI}
\usage{
dragSetUI(id, textval = c(""))
}
\arguments{
\item{id}{The div id of the draggable element set}
}
\value{
dragable div element
}
\description{
dragSetUI
}

