% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropUI.R
\name{dropUI}
\alias{dropUI}
\title{dropUI}
\usage{
dropUI(id, style = NULL, class = "dropelement", row_n = 1, col_n = 1)
}
\arguments{
\item{id}{The div id of the element}

\item{style}{The css styling}

\item{class}{The css class style}

\item{row_n}{set the number of rows to be used in the vertical selection. Height is proportional to the number of rows.}

\item{col_n}{set the number of columns to be used in the vertical selection. Width is proportional to the number of columns.}
}
\value{
div element with nested elements if vertical is true
}
\description{
dropUI
}

