% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDND.R
\docType{package}
\name{shinyDND}
\alias{shinyDND}
\alias{shinyDND-package}
\title{ShinyDND: A package for creating drag and drop elements in Shiny.}
\description{
The ShinyDND package provides three categories of important functions:
dragUI, dropUI, and dragSetUI.
}
\section{ShinyDND functions}{

The ShinyDND functions ...
}

